Troubles ....

+++ Detectors/HMPID/reconstruction/include/HMPIDReconstruction/HmpidDecoder.h
@@ -155,7 +155,7 @@ class HmpidDecoder
     HmpidEquipment* evaluateHeaderContents(int EquipmentIndex);
     void updateStatistics(HmpidEquipment *eq);
   protected:
-    virtual void setPad(HmpidEquipment *eq, int col, int dil, int ch, int charge) = 0;
+    virtual void setPad(HmpidEquipment *eq, int col, int dil, int ch, uint16_t charge) = 0;
     virtual bool getBlockFromStream(uint32_t **streamPtr, uint32_t Size) = 0;
     virtual bool getHeaderFromStream(uint32_t **streamPtr) = 0;
     virtual bool getWordFromStream(uint32_t *word) = 0;
